+++ src/pages/Admin/LearningPath/LearningPath.jsx
@@ -5,44 +5,13 @@ import { Link } from "react-router-dom";
 import Table from "../../../components/Table/Table";
 import { AiFillDelete } from "react-icons/ai";
 import { MdModeEdit } from "react-icons/md";
-import ReactPaginate from "react-paginate";
 import ModalAlert from "../../../components/ModalAlert/ModalAlert";
+import Pagination from "../../../components/Pagination/Pagination";
 
 const LearningPath = () => {
   const [openModal, setOpenModal] = useState(false);
   const [elementSeleted, setElementSeleted] = useState(null);
-  const [data, setData] = useState(LearningPaths);
-  const [pageNumber, setPageNumber] = useState(0);
 
-  const dataPerPage = 10;
-  const pagesVisited = pageNumber * dataPerPage;
-
-  const displayData = data
-    .slice(pagesVisited, pagesVisited + dataPerPage)
-    .map(({ title, description, link, imagen }) => (
-      <>
-        <li>{title}</li>
-        <li className="TextClipped">{description}</li>
-        <li className="TextClipped">{link}</li>
-        <li className="TextClipped">{imagen}</li>
-        <li className="Table__actions">
-          <button onClick={() => handleOpenModal(title)}>
-            <AiFillDelete className="BtnDelete" />
-          </button>
-          <Link to="/">
-            <button>
-              <MdModeEdit />
-            </button>
-          </Link>
-        </li>
-      </>
-    ));
-  const pageCount = Math.ceil(data.length / dataPerPage);
-
-  const changePage = ({ selected }) => {
-    setPageNumber(selected);
-  };
-  
   const handleOpenModal = (id) => {
     setOpenModal((prevState) => !prevState);
     setElementSeleted(id);
@@ -68,18 +37,26 @@ const LearningPath = () => {
           "Acciones",
         ]}
       >
-        {displayData}
-        <ReactPaginate
-          previousLabel="Atras"
-          nextLabel="Siguiente"
-          pageCount={pageCount}
-          onPageChange={changePage}
-          containerClassName="ButtonsPagination"
-          previousLinkClassName="BtnAtras"
-          nextLinkClassName="BtnSiguiente"
-          disabledClassName="PaginationDisabled"
-          activeClassName="PaginationActive"
-        />
+        <Pagination data={LearningPaths}>
+          {LearningPaths.map(({ title, description, link, imagen }) => (
+            <>
+              <li>{title}</li>
+              <li className="TextClipped">{description}</li>
+              <li className="TextClipped">{link}</li>
+              <li className="TextClipped">{imagen}</li>
+              <li className="Table__actions">
+                <button onClick={() => handleOpenModal(title)}>
+                  <AiFillDelete className="BtnDelete" />
+                </button>
+                <Link to="/">
+                  <button>
+                    <MdModeEdit />
+                  </button>
+                </Link>
+              </li>
+            </>
+          ))}
+        </Pagination>
       </Table>
       {openModal && (
         <ModalAlert
