+++ src/pages/Admin/Resources/Resources.jsx
@@ -5,43 +5,12 @@ import { RecommendedResources } from "../../../assets/mocks";
 import { Link } from "react-router-dom";
 import { AiFillDelete } from "react-icons/ai";
 import { MdModeEdit } from "react-icons/md";
-import ReactPaginate from "react-paginate";
 import ModalAlert from "../../../components/ModalAlert/ModalAlert";
+import Pagination from "../../../components/Pagination/Pagination";
 
 const Resources = () => {
   const [openModal, setOpenModal] = useState(false);
   const [elementSeleted, setElementSeleted] = useState(null);
-  const [data, setData] = useState(RecommendedResources)
-  const [pageNumber, setPageNumber] = useState(0);
-
-  const dataPerPage = 10;
-  const pagesVisited = pageNumber * dataPerPage;
-
-  const displayData = data
-    .slice(pagesVisited, pagesVisited + dataPerPage)
-    .map(({ title, link, type }) => (
-      <>
-        <li>{title}</li>
-        <li className="TextClipped">{link}</li>
-        <li>{type}</li>
-        <li className="Table__actions">
-          <button onClick={() => handleOpenModal(title)}>
-            <AiFillDelete className="BtnDelete" />
-          </button>
-          <Link to="/">
-            <button>
-              <MdModeEdit />
-            </button>
-          </Link>
-        </li>
-      </>
-    ))
-
-  const pageCount = Math.ceil(data.length / dataPerPage);
-
-  const changePage = ({ selected }) => {
-    setPageNumber(selected);
-  };
 
   const handleOpenModal = (id) => {
     setOpenModal((prevState) => !prevState);
@@ -62,18 +31,25 @@ const Resources = () => {
       </Link>
 
       <Table headers={["Titulo", "Link", "Tipo", " Acciones"]}>
-        {displayData}
-        <ReactPaginate
-          previousLabel="Atras"
-          nextLabel="Siguiente"
-          pageCount={pageCount}
-          onPageChange={changePage}
-          containerClassName="ButtonsPagination"
-          previousLinkClassName="BtnAtras"
-          nextLinkClassName="BtnSiguiente"
-          disabledClassName="PaginationDisabled"
-          activeClassName="PaginationActive"
-        />
+        <Pagination data={RecommendedResources}>
+          {RecommendedResources.map(({ title, link, type }) => (
+            <>
+              <li>{title}</li>
+              <li className="TextClipped">{link}</li>
+              <li>{type}</li>
+              <li className="Table__actions">
+                <button onClick={() => handleOpenModal(title)}>
+                  <AiFillDelete className="BtnDelete" />
+                </button>
+                <Link to="/">
+                  <button>
+                    <MdModeEdit />
+                  </button>
+                </Link>
+              </li>
+            </>
+          ))}
+        </Pagination>
       </Table>
       {openModal && (
         <ModalAlert
